+++ src/endpoints/minidump.rs
@@ -4,7 +4,7 @@ use axum::response::IntoResponse;
 
 use crate::endpoints::{
     authenticate_and_prefilter, check_event_filter, error_response, overloaded_response,
-    sentry_response,
+    sentry_response, sentry_response_with_discarded,
 };
 use crate::models::StorableAttachment;
 use crate::server::AppState;
@@ -72,7 +72,7 @@ pub async fn handle(
     crate::enrich::enrich_event(&mut event);
 
     if check_event_filter(&state, &event, project_id) {
-        return sentry_response(&event_id).into_response();
+        return sentry_response_with_discarded(&event_id, 1).into_response();
     }
 
     if state